fix: Moved svg to png to a util function
